+++ CMakeLists.txt
@@ -54,6 +54,7 @@ add_library ( android-vulkan
     app/src/main/cpp/sources/renderer.cpp
     app/src/main/cpp/sources/vulkan_utils.cpp
     app/src/main/cpp/sources/GXCommon/GXMath.cpp
+    app/src/main/cpp/sources/GXCommon/Vulkan/GXMathBackend.cpp
     app/src/main/cpp/sources/mandelbrot/mandelbrot_analytic_color.cpp
     app/src/main/cpp/sources/mandelbrot/mandelbrot_base.cpp
     app/src/main/cpp/sources/mandelbrot/mandelbrot_lut_color.cpp
